@@ -1,19 +1,23 @@
 package app.droidinfo.ui;
 
+import android.Manifest;
 import android.app.Activity;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
 import android.content.SharedPreferences;
+import android.content.pm.PackageManager;
 import android.graphics.Typeface;
 import android.opengl.GLSurfaceView;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
+import android.os.Process;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.Window;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import java.util.Timer;
 import java.util.TimerTask;
@@ -21,6 +25,8 @@ import java.util.TimerTask;
 import javax.microedition.khronos.egl.EGLConfig;
 import javax.microedition.khronos.opengles.GL10;
 
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
 import app.droidinfo.R;
 import app.droidinfo.helper.BatteryHelper;
 import app.droidinfo.helper.DisplayHelper;
@@ -73,10 +79,15 @@ public class SplashActivity extends AppCompatActivity implements GLSurfaceView.R
                 .putString("RESOLUTION", DisplayHelper.getResolution(SplashActivity.this))
                 .apply();
 
-        Timer timer = new Timer();
-        timer.schedule(new TimerTask() {
-            @Override
-            public void run() {
+        if (ContextCompat.checkSelfPermission(this, Manifest.permission.READ_PHONE_STATE)
+                != PackageManager.PERMISSION_GRANTED) {
+            ActivityCompat.requestPermissions(this,
+                    new String[]{Manifest.permission.READ_PHONE_STATE}, 0);
+        } else if (ContextCompat.checkSelfPermission(this, Manifest.permission.READ_PHONE_STATE) == PackageManager.PERMISSION_GRANTED){
+            Timer timer = new Timer();
+            timer.schedule(new TimerTask() {
+                @Override
+                public void run() {
                 /*boolean firstRun = sharedPreferences.getBoolean("FIRST_RUN", true);
                 if (firstRun) {
                     startActivity(new Intent(SplashActivity.this, IntroActivity.class));
@@ -85,10 +96,47 @@ public class SplashActivity extends AppCompatActivity implements GLSurfaceView.R
                     startActivity(new Intent(SplashActivity.this, MainActivity.class));
                     finish();
                 }*/
-                startActivity(new Intent(SplashActivity.this, MainActivity.class));
-                finish();
+                    startActivity(new Intent(SplashActivity.this, MainActivity.class));
+                    finish();
+                }
+            }, 2000);
+        }
+
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, String permissions[], int[] grantResults) {
+        switch (requestCode) {
+            case 0: {
+                // If request is cancelled, the result arrays are empty.
+                if (grantResults.length > 0
+                        && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                    Timer timer = new Timer();
+                    timer.schedule(new TimerTask() {
+                        @Override
+                        public void run() {
+                            /*boolean firstRun = sharedPreferences.getBoolean("FIRST_RUN", true);
+                            if (firstRun) {
+                                startActivity(new Intent(SplashActivity.this, IntroActivity.class));
+                                finish();
+                            } else {
+                                startActivity(new Intent(SplashActivity.this, MainActivity.class));
+                                finish();
+                            }*/
+                            startActivity(new Intent(SplashActivity.this, MainActivity.class));
+                            finish();
+                        }
+                    }, 2000);
+                } else {
+                    Toast.makeText(this, "Soory, you can't use app.", Toast.LENGTH_SHORT).show();
+                    android.os.Process.killProcess(Process.myPid());
+                }
+                return;
             }
-        }, 2000);
+
+            // other 'case' lines to check for other
+            // permissions this app might request.
+        }
     }
 
     @Override
